Create dedicated Rfc003Actor
This would facilitate the assert code as the cnd REST API is different
for rfc003 and the new split protocols (han, herc20, halight).
