+++ api_tests/src/actors/actor.ts
@@ -1,31 +1,24 @@
 import {
     Cnd,
-    ComitClient,
     Entity,
     HanEthereumEtherHalightLightningBitcoinRequestBody,
     LedgerAction,
     Swap,
     SwapDetails,
     Transaction,
-    TransactionStatus,
     Wallets as SdkWallets,
 } from "comit-sdk";
 import { Logger } from "log4js";
 import { E2ETestActorConfig } from "../config";
-import { Asset, AssetKind, toKey, toKind } from "../asset";
+import { Asset, AssetKind, toKey } from "../asset";
 import { CndInstance } from "../cnd/cnd_instance";
 import { Ledger, LedgerKind } from "../ledgers/ledger";
 import { LedgerConfig, sleep } from "../utils";
-import { Wallet, Wallets } from "../wallets";
 import { Actors } from "./index";
+import { Wallets } from "../wallets";
 import { sha256 } from "js-sha256";
 import { InvoiceState } from "@radar/lnrpc";
 import { defaultLedgerDescriptionForLedger } from "./defaults";
-import {
-    defaultAssetDescription,
-    defaultExpiryTimes,
-    defaultLedgerKindForAsset,
-} from "./rfc003_defaults";
 
 export type ActorNames = "alice" | "bob" | "charlie";
 
@@ -65,22 +58,21 @@ export class Actor {
     public actors: Actors;
     public wallets: Wallets;
 
-    private comitClient: ComitClient;
     readonly cnd: Cnd;
-    private swap: Swap;
+    public swap: Swap;
 
-    private alphaLedger: Ledger;
-    private alphaAsset: Asset;
+    public alphaLedger: Ledger;
+    public alphaAsset: Asset;
 
-    private betaLedger: Ledger;
-    private betaAsset: Asset;
+    public betaLedger: Ledger;
+    public betaAsset: Asset;
 
-    private readonly startingBalances: Map<string, bigint>;
-    private readonly expectedBalanceChanges: Map<string, bigint>;
+    public readonly startingBalances: Map<string, bigint>;
+    public readonly expectedBalanceChanges: Map<string, bigint>;
 
     constructor(
         public readonly logger: Logger,
-        private readonly cndInstance: CndInstance,
+        public readonly cndInstance: CndInstance,
         private readonly name: ActorNames
     ) {
         this.wallets = new Wallets({});
@@ -89,152 +81,6 @@ export class Actor {
 
         this.startingBalances = new Map();
         this.expectedBalanceChanges = new Map();
-    }
-
-    public async sendRequest(
-        maybeAlpha?: AssetKind | { ledger: LedgerKind; asset: AssetKind },
-        maybeBeta?: AssetKind | { ledger: LedgerKind; asset: AssetKind }
-    ) {
-        this.logger.info("Sending swap request");
-
-        // By default, we will send the swap request to bob
-        const to = this.actors.bob;
-
-        let alphaAssetKind: AssetKind;
-        let alphaLedgerKind: LedgerKind;
-        if (!maybeAlpha) {
-            alphaAssetKind = this.defaultAlphaAssetKind();
-            alphaLedgerKind = this.defaultAlphaLedgerKind();
-        } else if (typeof maybeAlpha === "string") {
-            alphaAssetKind = maybeAlpha;
-            alphaLedgerKind = defaultLedgerKindForAsset(alphaAssetKind);
-        } else {
-            alphaAssetKind = maybeAlpha.asset;
-            alphaLedgerKind = maybeAlpha.ledger;
-        }
-
-        this.alphaLedger = defaultLedgerDescriptionForLedger(alphaLedgerKind);
-        this.alphaAsset = defaultAssetDescription(
-            alphaAssetKind,
-            alphaLedgerKind
-        );
-        to.alphaLedger = this.alphaLedger;
-        to.alphaAsset = this.alphaAsset;
-
-        this.logger.debug(
-            "Derived Alpha Ledger %o from %s",
-            this.alphaLedger,
-            alphaLedgerKind
-        );
-        this.logger.debug(
-            "Derived Alpha Asset %o from %s",
-            this.alphaAsset,
-            alphaAssetKind
-        );
-
-        let betaAssetKind;
-        let betaLedgerKind;
-        if (!maybeBeta) {
-            betaAssetKind = this.defaultBetaAssetKind();
-            betaLedgerKind = this.defaultBetaLedgerKind();
-        } else if (typeof maybeBeta === "string") {
-            betaAssetKind = maybeBeta;
-            betaLedgerKind = defaultLedgerKindForAsset(betaAssetKind);
-        } else {
-            betaAssetKind = maybeBeta.asset;
-            betaLedgerKind = maybeBeta.ledger;
-        }
-
-        this.betaLedger = defaultLedgerDescriptionForLedger(betaLedgerKind);
-        this.betaAsset = defaultAssetDescription(betaAssetKind, betaLedgerKind);
-        to.betaLedger = this.betaLedger;
-        to.betaAsset = this.betaAsset;
-
-        this.logger.debug(
-            "Derived Beta Ledger %o from %s",
-            this.betaLedger,
-            betaLedgerKind
-        );
-        this.logger.debug(
-            "Derived Beta Asset %o from %s",
-            this.betaAsset,
-            betaAssetKind
-        );
-
-        const listPromises: Promise<void>[] = [
-            this.initializeDependencies(),
-            to.initializeDependencies(),
-        ];
-        await Promise.all(listPromises);
-
-        await this.setStartingBalance([
-            this.alphaAsset,
-            {
-                name: this.betaAsset.name,
-                ledger: this.betaLedger.name,
-                quantity: "0",
-            },
-        ]);
-        await to.setStartingBalance([
-            {
-                name: to.alphaAsset.name,
-                ledger: this.alphaLedger.name,
-                quantity: "0",
-            },
-            to.betaAsset,
-        ]);
-
-        this.expectedBalanceChanges.set(
-            toKey(this.betaAsset),
-            BigInt(this.betaAsset.quantity)
-        );
-        to.expectedBalanceChanges.set(
-            toKey(this.alphaAsset),
-            BigInt(to.alphaAsset.quantity)
-        );
-
-        const isLightning =
-            this.alphaLedger.name === "lightning" ||
-            this.betaLedger.name === "lightning";
-
-        if (isLightning) {
-            this.logger.debug("Using lightning routes on cnd REST API");
-            return;
-        }
-
-        const comitClient: ComitClient = this.getComitClient();
-
-        const payload = {
-            alpha_ledger: this.alphaLedger,
-            beta_ledger: this.betaLedger,
-            alpha_asset: this.alphaAsset,
-            beta_asset: this.betaAsset,
-            peer: {
-                peer_id: await to.cnd.getPeerId(),
-                address_hint: await to.cnd
-                    .getPeerListenAddresses()
-                    .then((addresses) => addresses[0]),
-            },
-            ...(await this.additionalIdentities(alphaAssetKind, betaAssetKind)),
-            ...defaultExpiryTimes(),
-        };
-
-        this.swap = await comitClient.sendSwap(payload);
-        to.swap = new Swap(to.cnd, this.swap.self, {
-            bitcoin: to.wallets.bitcoin.inner,
-            ethereum: to.wallets.ethereum.inner,
-        });
-        this.logger.debug("Created new swap at %s", this.swap.self);
-
-        return this.swap;
-    }
-
-    public async accept() {
-        if (!this.swap) {
-            throw new Error("Cannot accept non-existent swap");
-        }
-
-        await this.swap.accept(Actor.defaultActionConfig);
     }
 
     public async createSwap(
@@ -311,46 +157,6 @@ export class Actor {
         );
     }
 
-    public async deploy() {
-        if (!this.swap) {
-            throw new Error("Cannot deploy htlc for nonexistent swap");
-        }
-
-        const transaction = await this.swap.deploy(Actor.defaultActionConfig);
-        let transactionId;
-        if (transaction instanceof Transaction) {
-            const status = await transaction.status(1);
-            transactionId = transaction.id;
-            if (status === TransactionStatus.Failed) {
-                throw new Error(`Transaction ${transactionId} failed`);
-            }
-        } else {
-            transactionId = transaction;
-        }
-
-        this.logger.debug(
-            "Deployed htlc for swap %s in %s",
-            this.swap.self,
-            transactionId
-        );
-
-        const entity = await this.swap.fetchDetails();
-        switch (entity.properties.role) {
-            case "Alice":
-                await this.actors.alice.assertAlphaDeployed();
-                if (this.actors.bob.cndInstance.isRunning()) {
-                    await this.actors.bob.assertAlphaDeployed();
-                }
-                break;
-            case "Bob":
-                if (this.actors.alice.cndInstance.isRunning()) {
-                    await this.actors.alice.assertBetaDeployed();
-                }
-                await this.actors.bob.assertBetaDeployed();
-                break;
-        }
-    }
-
     public async init(config?: {
         maxTimeoutSecs: number;
         tryIntervalSecs: number;
@@ -401,110 +207,6 @@ export class Actor {
         }
     }
 
-    public async fundLowGas(hexGasLimit: string) {
-        const response = await this.swap.tryExecuteSirenAction<LedgerAction>(
-            "fund",
-            {
-                maxTimeoutSecs: 10,
-                tryIntervalSecs: 1,
-            }
-        );
-        response.data.payload.gas_limit = hexGasLimit;
-        const transaction = await this.swap.doLedgerAction(response.data);
-        if (transaction instanceof Transaction) {
-            const status = await transaction.status(1);
-            if (status !== TransactionStatus.Failed) {
-                throw new Error(
-                    "Deploy with low gas transaction was successful."
-                );
-            }
-        } else {
-            throw new Error(
-                "Internal error: Transaction class expected for Ethereum."
-            );
-        }
-    }
-
-    public async overfund() {
-        const response = await this.swap.tryExecuteSirenAction<LedgerAction>(
-            "fund",
-            {
-                maxTimeoutSecs: 10,
-                tryIntervalSecs: 1,
-            }
-        );
-        const amount = response.data.payload.amount;
-        const overfundAmount = amount * 1.01;
-
-        response.data.payload.amount = overfundAmount;
-
-        const txid = await this.swap.doLedgerAction(response.data);
-        this.logger.debug(
-            "Overfunded swap %s in %s with %d instead of %d",
-            this.swap.self,
-            txid,
-            overfundAmount,
-            amount
-        );
-    }
-
-    public async underfund() {
-        const response = await this.swap.tryExecuteSirenAction<LedgerAction>(
-            "fund",
-            {
-                maxTimeoutSecs: 10,
-                tryIntervalSecs: 1,
-            }
-        );
-        const amount = response.data.payload.amount;
-        const underfundAmount = amount * 0.01;
-
-        response.data.payload.amount = underfundAmount;
-
-        const txid = await this.swap.doLedgerAction(response.data);
-        this.logger.debug(
-            "Underfunded swap %s in %s with %d instead of %d",
-            this.swap.self,
-            txid,
-            underfundAmount,
-            amount
-        );
-    }
-
-    public async refund() {
-        if (!this.swap) {
-            throw new Error("Cannot refund non-existent swap");
-        }
-
-        const role = await this.whoAmI();
-        switch (role) {
-            case "Alice":
-                await this.waitForAlphaExpiry();
-                break;
-            case "Bob":
-                await this.waitForBetaExpiry();
-                break;
-        }
-
-        const txid = await this.swap.refund(Actor.defaultActionConfig);
-        this.logger.debug("Refunded swap %s in %s", this.swap.self, txid);
-
-        switch (role) {
-            case "Alice":
-                await this.actors.alice.assertAlphaRefunded();
-                if (this.actors.bob.cndInstance.isRunning()) {
-                    await this.actors.bob.assertAlphaRefunded();
-                }
-                break;
-            case "Bob":
-                if (this.actors.alice.cndInstance.isRunning()) {
-                    await this.actors.alice.assertBetaRefunded();
-                }
-                await this.actors.bob.assertBetaRefunded();
-                break;
-        }
-    }
-
     public async redeem() {
         if (!this.swap) {
             throw new Error("Cannot redeem non-existent swap");
@@ -530,28 +232,6 @@ export class Actor {
         }
     }
 
-    public async redeemWithHighFee() {
-        // Hack the bitcoin fee per WU returned by the wallet
-        this.wallets.bitcoin.inner.getFee = () => "100000000";
-
-        return this.swap.tryExecuteSirenAction<LedgerAction>("redeem", {
-            maxTimeoutSecs: 10,
-            tryIntervalSecs: 1,
-        });
-    }
-
-    public async currentSwapIsAccepted() {
-        let swapEntity;
-
-        do {
-            swapEntity = await this.swap.fetchDetails();
-
-            await sleep(200);
-        } while (
-            swapEntity.properties.state.communication.status !== "ACCEPTED"
-        );
-    }
-
     public async assertSwapped() {
         this.logger.debug("Checking if cnd reports status 'SWAPPED'");
 
@@ -567,114 +247,23 @@ export class Actor {
     }
 
     public async assertBalances() {
-        for (const [
-            assetKey,
-            expectedBalanceChange,
-        ] of this.expectedBalanceChanges.entries()) {
-            this.logger.debug(
-                "Checking that %s balance changed by %d",
-                assetKey,
-                expectedBalanceChange
-            );
-
-            const { asset, ledger } = toKind(assetKey);
-
-            const wallet = this.wallets[ledger];
-            const expectedBalance =
-                this.startingBalances.get(assetKey) + expectedBalanceChange;
-            const maximumFee = BigInt(wallet.MaximumFee);
-
-            const balanceInclFees = expectedBalance - maximumFee;
-
-            const currentWalletBalance = await wallet.getBalanceByAsset(
-                defaultAssetDescription(asset, ledger)
-            );
-            expect(currentWalletBalance).toBeGreaterThanOrEqual(
-                // @ts-ignore: Jest supports bigint, types to be fixed updated with
-                // https://github.com/DefinitelyTyped/DefinitelyTyped/pull/44368
-                balanceInclFees
-            );
-
-            this.logger.debug(
-                "Balance check was positive, current balance is %d",
-                currentWalletBalance
-            );
-        }
-    }
-
-    public async assertRefunded() {
-        this.logger.debug("Checking if swap @ %s was refunded", this.swap.self);
-
-        for (const [assetKey] of this.startingBalances.entries()) {
-            const { asset, ledger } = toKind(assetKey);
-
-            const wallet = this.wallets[ledger];
-            const maximumFee = BigInt(wallet.MaximumFee);
-
-            this.logger.debug(
-                "Checking that %s balance changed by max %d (MaximumFee)",
-                assetKey,
-                maximumFee
-            );
-            const expectedBalance = this.startingBalances.get(assetKey);
-            const currentWalletBalance = await wallet.getBalanceByAsset(
-                defaultAssetDescription(asset, ledger)
-            );
-            const balanceInclFees = expectedBalance - maximumFee;
-            expect(currentWalletBalance).toBeGreaterThanOrEqual(
-                // @ts-ignore: Jest supports bigint, types to be fixed updated with
-                // https://github.com/DefinitelyTyped/DefinitelyTyped/pull/44368
-                balanceInclFees
-            );
-        }
-    }
-
-    public async assertAlphaDeployed() {
-        await this.assertLedgerState("alpha_ledger", "DEPLOYED");
-    }
-
-    public async assertBetaDeployed() {
-        await this.assertLedgerState("beta_ledger", "DEPLOYED");
+        // TODO: WTF is this checking for halight?
     }
 
     public async assertAlphaFunded() {
-        await this.assertLedgerState("alpha_ledger", "FUNDED");
+        // TODO: Actually assert
     }
 
     public async assertBetaFunded() {
-        await this.assertLedgerState("beta_ledger", "FUNDED");
+        // TODO: Actually assert
     }
 
     public async assertAlphaRedeemed() {
-        await this.assertLedgerState("alpha_ledger", "REDEEMED");
+        // TODO: Actually assert
     }
 
     public async assertBetaRedeemed() {
-        await this.assertLedgerState("beta_ledger", "REDEEMED");
-    }
-
-    public async assertAlphaRefunded() {
-        await this.assertLedgerState("alpha_ledger", "REFUNDED");
-    }
-
-    public async assertBetaRefunded() {
-        await this.assertLedgerState("beta_ledger", "REFUNDED");
-    }
-
-    public async assertAlphaIncorrectlyFunded() {
-        await this.assertLedgerState("alpha_ledger", "INCORRECTLY_FUNDED");
-    }
-
-    public async assertBetaIncorrectlyFunded() {
-        await this.assertLedgerState("beta_ledger", "INCORRECTLY_FUNDED");
-    }
-
-    public async assertAlphaNotDeployed() {
-        await this.assertLedgerState("alpha_ledger", "NOT_DEPLOYED");
-    }
-
-    public async assertBetaNotDeployed() {
-        await this.assertLedgerState("beta_ledger", "NOT_DEPLOYED");
+        // TODO: Actually assert
     }
 
     public async start() {
@@ -686,34 +275,12 @@ export class Actor {
         this.cndInstance.stop();
     }
 
-    public async restart() {
-        await this.stop();
-        await this.start();
-    }
-
     public async dumpState() {
-        this.logger.debug("dumping current state");
-
-        if (this.swap) {
-            const swapDetails = await this.swap.fetchDetails();
-
-            this.logger.debug("swap status: %s", swapDetails.properties.status);
-            this.logger.debug("swap details: ", JSON.stringify(swapDetails));
-
-            this.logger.debug(
-                "alpha ledger wallet balance %d",
-                await this.alphaLedgerWallet().getBalanceByAsset(
-                    this.alphaAsset
-                )
-            );
-            this.logger.debug(
-                "beta ledger wallet balance %d",
-                await this.betaLedgerWallet().getBalanceByAsset(this.betaAsset)
-            );
-        }
+        // TODO: Actually dump a split protocol state
     }
 
-    public async whoAmI() {
+    // TODO: Check if this correct or if we can use getName
+    public async whoAmI(): Promise<string> {
         const entity = await this.swap.fetchDetails();
         return entity.properties.role;
     }
@@ -722,152 +289,7 @@ export class Actor {
         return this.name;
     }
 
-    private async waitForAlphaExpiry() {
-        const swapDetails = await this.swap.fetchDetails();
-
-        const expiry = swapDetails.properties.state.communication.alpha_expiry;
-        const wallet = this.alphaLedgerWallet();
-
-        await this.waitForExpiry(wallet, expiry);
-    }
-
-    private async waitForBetaExpiry() {
-        const swapDetails = await this.swap.fetchDetails();
-
-        const expiry = swapDetails.properties.state.communication.beta_expiry;
-        const wallet = this.betaLedgerWallet();
-
-        await this.waitForExpiry(wallet, expiry);
-    }
-
-    private alphaLedgerWallet() {
-        return this.wallets.getWalletForLedger(this.alphaLedger.name);
-    }
-
-    private betaLedgerWallet() {
-        return this.wallets.getWalletForLedger(this.betaLedger.name);
-    }
-
-    private async waitForExpiry(wallet: Wallet, expiry: number) {
-        let currentBlockchainTime = await wallet.getBlockchainTime();
-
-        this.logger.debug(
-            `Current blockchain time is ${currentBlockchainTime}`
-        );
-
-        let diff = expiry - currentBlockchainTime;
-
-        if (diff > 0) {
-            this.logger.debug(`Waiting for blockchain time to pass ${expiry}`);
-
-            while (diff > 0) {
-                await sleep(1000);
-
-                currentBlockchainTime = await wallet.getBlockchainTime();
-                diff = expiry - currentBlockchainTime;
-
-                this.logger.debug(
-                    `Current blockchain time is ${currentBlockchainTime}`
-                );
-            }
-        }
-    }
-
-    private async assertLedgerState(
-        ledger: "alpha_ledger" | "beta_ledger",
-        status:
-            | "NOT_DEPLOYED"
-            | "DEPLOYED"
-            | "FUNDED"
-            | "REDEEMED"
-            | "REFUNDED"
-            | "INCORRECTLY_FUNDED"
-    ) {
-        // This is hackery of the higest order.
-        if (
-            this.betaLedger.name === LedgerKind.Lightning ||
-            this.alphaLedger.name === LedgerKind.Lightning
-        ) {
-            return;
-        }
-
-        this.logger.debug(
-            "Waiting for cnd to see %s in state %s for swap @ %s",
-            ledger,
-            status,
-            this.swap.self
-        );
-
-        let swapEntity;
-
-        do {
-            swapEntity = await this.swap.fetchDetails();
-
-            await sleep(200);
-        } while (swapEntity.properties.state[ledger].status !== status);
-
-        this.logger.debug(
-            "cnd saw %s in state %s for swap @ %s",
-            ledger,
-            status,
-            this.swap.self
-        );
-    }
-
-    private async additionalIdentities(
-        alphaAsset: AssetKind,
-        betaAsset: AssetKind
-    ) {
-        if (alphaAsset === "bitcoin" && betaAsset === "ether") {
-            return {
-                beta_ledger_redeem_identity: this.wallets.ethereum.account(),
-            };
-        }
-
-        return {};
-    }
-
-    private async initializeDependencies() {
-        const lightningNeeded =
-            this.alphaLedger.name === "lightning" ||
-            this.betaLedger.name === "lightning";
-
-        const walletPromises: Promise<void>[] = [];
-        for (const ledgerName of [
-            this.alphaLedger.name,
-            this.betaLedger.name,
-        ]) {
-            walletPromises.push(
-                this.wallets.initializeForLedger(
-                    ledgerName,
-                    this.logger,
-                    this.name
-                )
-            );
-        }
-
-        await Promise.all(walletPromises);
-
-        if (!lightningNeeded) {
-            this.comitClient = new ComitClient(this.cnd)
-                .withBitcoinWallet(
-                    this.wallets.getWalletForLedger("bitcoin").inner
-                )
-                .withEthereumWallet(
-                    this.wallets.getWalletForLedger("ethereum").inner
-                );
-        }
-    }
-
-    private getComitClient(): ComitClient {
-        if (!this.comitClient) {
-            throw new Error("ComitClient is not initialised");
-        }
-
-        return this.comitClient;
-    }
-
-    private async setStartingBalance(assets: Asset[]) {
+    public async setStartingBalance(assets: Asset[]) {
         for (const asset of assets) {
             if (parseFloat(asset.quantity) === 0) {
                 this.startingBalances.set(toKey(asset), BigInt(0));
@@ -897,46 +319,6 @@ export class Actor {
         }
     }
 
-    private defaultAlphaAssetKind() {
-        const defaultAlphaAssetKind = AssetKind.Bitcoin;
-        this.logger.info(
-            "AssetKind for alpha asset not specified, defaulting to %s",
-            defaultAlphaAssetKind
-        );
-
-        return defaultAlphaAssetKind;
-    }
-
-    private defaultAlphaLedgerKind() {
-        const defaultAlphaLedgerKind = LedgerKind.Bitcoin;
-        this.logger.info(
-            "LedgerKind for alpha ledger not specified, defaulting to %s",
-            defaultAlphaLedgerKind
-        );
-
-        return defaultAlphaLedgerKind;
-    }
-
-    private defaultBetaAssetKind() {
-        const defaultBetaAssetKind = AssetKind.Ether;
-        this.logger.info(
-            "AssetKind for beta asset not specified, defaulting to %s",
-            defaultBetaAssetKind
-        );
-
-        return defaultBetaAssetKind;
-    }
-
-    private defaultBetaLedgerKind() {
-        const defaultBetaLedgerKind = LedgerKind.Ethereum;
-        this.logger.info(
-            "LedgerKind for beta ledger not specified, defaulting to %s",
-            defaultBetaLedgerKind
-        );
-
-        return defaultBetaLedgerKind;
-    }
-
     public cndHttpApiUrl() {
         const socket = this.cndInstance.getConfigFile().http_api.socket;
         return `http://${socket}`;
@@ -944,6 +326,7 @@ export class Actor {
 
     public async pollCndUntil(
         location: string,
+        // TODO: Use the correct type
         predicate: (body: Entity) => boolean
     ): Promise<Entity> {
         const response = await this.cnd.fetch(location);
@@ -959,6 +342,7 @@ export class Actor {
         }
     }
 
+    // TODO: Use the correct type
     public async pollSwapDetails(
         swapUrl: string,
         iteration: number = 0
@@ -976,7 +360,7 @@ export class Actor {
         }
     }
 
-    /// This is to be removed once cnd supports lightning
+    /// TODO: Remove once all asserts are in place
     public lnCreateSha256Secret(): { secret: string; secretHash: string } {
         const secretBuf = Buffer.alloc(32);
         for (let i = 0; i < secretBuf.length; i++) {
