+++ api_tests/src/actors/rfc003_defaults.ts
@@ -17,17 +17,11 @@ export function defaultLedgerKindForAsset(asset: AssetKind): LedgerKind {
 }
 
 /**
- * WIP as the cnd REST API routes for lightning are not yet defined.
  * @param ledger
  * @returns The ledger formatted as needed for the request body to cnd HTTP API on the lightning route.
  */
 export function defaultLedgerDescriptionForLedger(ledger: LedgerKind): Ledger {
     switch (ledger) {
-        case LedgerKind.Lightning: {
-            return {
-                name: LedgerKind.Lightning,
-            };
-        }
         case LedgerKind.Bitcoin: {
             return {
                 name: LedgerKind.Bitcoin,
