+++ api_tests/tests/bitcoin_erc20.ts
@@ -5,11 +5,17 @@
 
 import { twoActorTest } from "../src/actor_test";
 import { AssetKind } from "../src/asset";
+import { Rfc003Actor } from "../src/actors/rfc003_actor";
 
 describe("E2E: Bitcoin/bitcoin - Ethereum/erc20", () => {
     it(
         "rfc003-btc-eth-erc20-alice-redeems-bob-redeems",
-        twoActorTest(async ({ alice, bob }) => {
+        twoActorTest(async (actors) => {
+            const [alice, bob] = Rfc003Actor.convert([
+                actors.alice,
+                actors.bob,
+            ]);
+
             await alice.sendRequest(AssetKind.Bitcoin, AssetKind.Erc20);
             await bob.accept();
 
@@ -27,7 +33,11 @@ describe("E2E: Bitcoin/bitcoin - Ethereum/erc20", () => {
 
     it(
         "rfc003-btc-eth-erc20-bob-refunds-alice-refunds",
-        twoActorTest(async ({ alice, bob }) => {
+        twoActorTest(async (actors) => {
+            const [alice, bob] = Rfc003Actor.convert([
+                actors.alice,
+                actors.bob,
+            ]);
             await alice.sendRequest(AssetKind.Bitcoin, AssetKind.Erc20);
             await bob.accept();
